requirements.txt
click==2.4
github3.py==0.9.0
six==1.7.3
